api-client/src/main/java/com/desk/java/apiclient/model/User.java
/*
 * Copyright (c) 2016, Salesforce.com, Inc.
 * All rights reserved.
 *
 * Redistribution and use in source and binary forms, with or without modification, are permitted provided
 * that the following conditions are met:
 *
 *    Redistributions of source code must retain the above copyright notice, this list of conditions and the
 *    following disclaimer.
 *
 *    Redistributions in binary form must reproduce the above copyright notice, this list of conditions and
 *    the following disclaimer in the documentation and/or other materials provided with the distribution.
 *
 *    Neither the name of Salesforce.com, Inc. nor the names of its contributors may be used to endorse or
 *    promote products derived from this software without specific prior written permission.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED
 * WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
 * PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
 * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
 * TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
 * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
 * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 * POSSIBILITY OF SUCH DAMAGE.
 */

package com.desk.java.apiclient.model;


import org.jetbrains.annotations.NotNull;
import org.jetbrains.annotations.Nullable;

import java.io.Serializable;

public class User implements Serializable {

    private static final long serialVersionUID = 752919141528370641L;

    private int id;
    private String name;
    private String publicName;
    private String email;
    private String level;
    private String avatar;
    private UserLinks _links;

    public int getId() {
        return id;
    }

    public void setId(int i) {
        this.id = i;
    }

    @NotNull
    public String getName() {
        return (name == null) ? "" : name;
    }

    public void setName(String n) {
        this.name = n;
    }

    @Nullable
    public String getPublicName() {
        return publicName;
    }

    public void setPublicName(String pn) {
        this.publicName = pn;
    }

    @Nullable
    public String getEmail() {
        return email;
    }

    public void setEmail(String e) {
        this.email = e;
    }

    @Nullable
    public String getLevel() {
        return level;
    }

    public void setLevel(String l) {
        this.level = l;
    }

    @NotNull
    public String getAvatar() {
        return (avatar == null) ? "" : avatar;
    }

    public void setAvatar(String a) {
        this.avatar = a;
    }

    @NotNull
    public UserLinks getLinks() {
        return _links == null ? _links = new UserLinks() : _links;
    }

    public void setLinks(UserLinks l) {
        this._links = l;
    }

    @Nullable
    public String getFiltersUrl() {
        if (getLinks().getFilters() == null) {
            return null;
        }
        return getLinks().getFilters().getUrl();
    }

    @NotNull
    public Link getSelfLink() {
        return getLinks().getSelf();
    }

    @Nullable
    public String getSelfLinkUrl() {
        return getSelfLink().getUrl();
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;

        User that = (User) o;

        return id == that.id;

    }

    @Override
    public int hashCode() {
        return id;
    }

}
